Agrego put y valido delete de usuarios

+++ backend/usuario.js
@@ -59,13 +59,47 @@ router.post("/",
   
 });
 
-router.put("/:id", (req, res) => {
+router.put("/:id",verificarAutenticacion, validarId,
+    body("nombre", "Nombre no valido")
+    .isAlpha("es-ES").isLength({max:20}),
+    body("apellido", "Apellido no valido")
+    .isAlpha("es-ES").isLength({max:20}),
+    body("email", "Email no valido")
+    .isEmail(),
+    verificarValidaciones, async (req, res) => {
+
+        const id = Numbers(req.params.id);
+
+        const [validar] = await db.execute("SELECT * FROM usuario WHERE id = ?", [id]);
+
+        if (validar.length === 0){
+            return res.status(404).json({success:false, message: "Usuario no encontrado"});
+        }
+        
+        const {nombre, apellido, email} = req.body;
+
+        const [validarExistencia] = await db.execute("SELECT * FROM usuario WHERE email = ? AND id != ?",[email,id]);
+
+        if (validarExistencia.length > 0){
+            return res.status(400).json({success: false, message: "Email ya registrado"});
+        }
+
+        const [result] = await db.execute("UPDATE usuario SET nombre = ?, apellido = ?, email = ? WHERE id = ?", [nombre, apellido, email, id]);
+
+        res.json({success: true, data: {id, nombre, apellido, email}});
 
 
 })
 
 router.delete("/:id", verificarAutenticacion, validarId, verificarValidaciones, async (req,res) => {
+
     const id = Number(req.params.id);
+
+    const [validar] = await db.execute("SELECT * FROM usuario WHERE id = ?", [id]);
+
+    if (validar.length === 0){
+        return res.status(404).json({success: false, message: "Usuario no encontrado"});
+    }
 
     await db.execute("DELETE FROM usuario WHERE id = ?", [id]);
